+++ docs/source/version_history.rst
@@ -6,7 +6,7 @@ development
 -----------
 
 - refactoring the Result-class
-- add workers as argument to start()
+- add worker argument to start()
 - fix: don't register crontasks twice